# backend/serverPanels/swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "title": "Nile Maret-Place Api",
        "description": "Description",
        "version": "1.0.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/earnings/{vendor_id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "vendor_id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/withdrawals/{vendor_id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "vendor_id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/getConversationMessages": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "customerId": {
                                    "example": "any"
                                },
                                "vendorId": {
                                    "example": "any"
                                },
                                "type": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/conversations": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "customerId": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/conversationsListforVendors": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "vendorId": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/sendInboxMessages": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "senderId": {
                                    "example": "any"
                                },
                                "recipientId": {
                                    "example": "any"
                                },
                                "content": {
                                    "example": "any"
                                },
                                "timestamp": {
                                    "example": "any"
                                },
                                "userType": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    }
}